double

+++ include/config.hpp
@@ -4,7 +4,7 @@
 static const int kBlockSizeCUDA = 256;
 
 // TODO: refactor everything into templates
-typedef float real;
-//typedef double real;
+//typedef float real;
+typedef double real;
 
 #endif
